# Conflicts:
#	app/src/main/java/uz/androbeck/virtualbank/ui/screens/bottom_nav_items/main/MainFragment.kt
#	app/src/main/res/navigation/main_nav_graph.xml
#	app/src/main/res/values-ru/strings.xml
#	app/src/main/res/values-uz/strings.xml
#	app/src/main/res/values/strings.xml
#	gradle/libs.versions.toml

+++ app/src/main/java/uz/androbeck/virtualbank/domain/ui_models/history/InComeAndOutComeUIModel.kt
@@ -4,7 +4,6 @@ data class InComeAndOutComeUIModel(
     val type: String? = null,
     val from: String? = null,
     val to: String? = null,
-    val amount: Float=0f,
-    val time: Long = 0,
-
-)
+    val amount: Long = 0,
+    val time: Long = 0
+    )
